Attempted exec-based solution
It doesn't work
